# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification System API",
        "description": "Event-driven notification system supporting SMS, Email, and Push channels with priority queues, rate limiting, retry logic, and idempotency.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/batches": {
            "post": {
                "tags": [
                    "Batches"
                ],
                "summary": "Create a batch of notifications",
                "description": "Creates up to 1000 notifications in a single request.",
                "operationId": "4ea2e8a3e2b899982d18be5f5ff50bfb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "notifications"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Flash sale campaign"
                                    },
                                    "metadata": {
                                        "type": "object"
                                    },
                                    "notifications": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "channel",
                                                "recipient",
                                                "content"
                                            ],
                                            "properties": {
                                                "channel": {
                                                    "type": "string",
                                                    "enum": [
                                                        "sms",
                                                        "email",
                                                        "push"
                                                    ]
                                                },
                                                "recipient": {
                                                    "type": "string"
                                                },
                                                "content": {
                                                    "type": "string"
                                                },
                                                "priority": {
                                                    "type": "string",
                                                    "enum": [
                                                        "high",
                                                        "normal",
                                                        "low"
                                                    ]
                                                },
                                                "metadata": {
                                                    "type": "object"
                                                },
                                                "scheduled_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "maxItems": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Batch created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "batch": {
                                                    "$ref": "#/components/schemas/NotificationBatchResource"
                                                },
                                                "created": {
                                                    "type": "integer"
                                                },
                                                "failed": {
                                                    "type": "integer"
                                                },
                                                "errors": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/batches/{id}": {
            "get": {
                "tags": [
                    "Batches"
                ],
                "summary": "Get a batch",
                "description": "Returns the batch status and counters by UUID.",
                "operationId": "c0a7cd7f431a32596621369243cf9b4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Batch UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/NotificationBatchResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotificationBatchResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed",
                            "partial_failed"
                        ]
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true
                    },
                    "total_count": {
                        "type": "integer"
                    },
                    "pending_count": {
                        "type": "integer"
                    },
                    "sent_count": {
                        "type": "integer"
                    },
                    "failed_count": {
                        "type": "integer"
                    },
                    "canceled_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}